requirements.txt
pandas
numpy
chromadb
sentence_transformers
argparse
langchain
langchain-core
langchain-ollama
langchain-groq
python-dotenv
